+++ Editor/NewExtendedEditorWindowPrompt/NewExtendedEditorWindowPrompt.cs
@@ -1,36 +1,62 @@
-﻿using System.Collections;
+﻿using System;
+using System.Collections;
+using System.IO;
+using System.Threading.Tasks;
 using UnityEditor;
 using ExtendedEditorWindows;
 using Unity.EditorCoroutines.Editor;
+using UnityEngine;
 
 public class NewExtendedEditorWindowPrompt : ExtendedEditorWindow<NewExtendedEditorWindowPrompt> {
         
     [MenuItem("Assets/Create/Extended Editor Window", false, 51)]
-    public static void Open() => OpenWindow("New Editor Window");
+    public static void Open() => OpenWindow("New Editor Window", true);
 
-    private string _windowTitle;
-    private string _windowFileName;
+    private bool _busy = false;
+
+    private string _windowTitle = "";
+    private string _windowFileName = "";
+    private string _windowMenuPath = "";
+
+    private Field<string> _titleField;
+    private Field<string> _menuPathField;
+    private Button _createButton;
+
+    private string _filePath;
 
     protected override void Initialize() {
         
-        var createEditorWindowButton = Button("CreateEditorWindow", button => {
-            button.enabled = false;
-            this.StartCoroutine(CreateDirectories());
-            this.StartCoroutine(CreateFiles());
-            button.enabled = true;
-        });
+        _createButton = Button("CreateEditorWindow", button => Generate());
+        _createButton.enabled = false;
 
-        createEditorWindowButton.enabled = false;
-        
-        Field("WindowTitle", "", field => {
+        _titleField = Field("Title", "", field => {
             _windowTitle = field.value;
             _windowFileName = field.value.Replace(" ", "");
-            createEditorWindowButton.enabled = field.value != "";
+            _filePath = $"Assets/Editor/Windows/{_windowFileName}/{_windowFileName}";
+        });
+        
+        _menuPathField = Field("MenuPath", "", field => {
+            _windowMenuPath = field.value;
         });
 
     }
 
-    private IEnumerator CreateDirectories() {
+    private void OnGUI() {
+        
+        _createButton.enabled = true;
+
+        if (_windowTitle == "") _createButton.enabled = false;
+        if (_windowMenuPath == "") _createButton.enabled = false;
+        if(_busy) _createButton.enabled = false;
+
+    }
+
+    private async void Generate() {
+        _busy = true;
+        await CreateDirectories();
+    }
+
+    private async Task CreateDirectories() {
         
         // Create the Editor folder if it doesn't exist.
         if (!AssetDatabase.IsValidFolder("Assets/Editor")) {
@@ -47,33 +73,60 @@ public class NewExtendedEditorWindowPrompt : ExtendedEditorWindow<NewExtendedEdi
             AssetDatabase.CreateFolder("Assets/Editor/Windows", _windowFileName);
         }
 
-        yield return null;
-        
+        await CreateFiles();
+
+    }
+    
+    private async Task CreateFiles() {
+
+        await GenerateScriptFile();
+        await GenerateStyleFile();
+        await GenerateTemplateFile();
+
+        _busy = false;
+
     }
 
-    private IEnumerator CreateFiles() {
+    private async Task GenerateScriptFile() {
         
-        // Generate the C# script file.
-        ProjectWindowUtil.CreateScriptAssetFromTemplateFile(
-            "Packages/com.sebastian-inman.extended-editor-windows/Templates~/ExtendedEditorWindow.cs.txt", 
-            $"Assets/Editor/Windows/{_windowFileName}/{_windowFileName}.cs"
-        );
-        
-        // Generate the USS stylesheet.
-        ProjectWindowUtil.CreateScriptAssetFromTemplateFile(
-            "Packages/com.sebastian-inman.extended-editor-windows/Templates~/ExtendedEditorWindow.uss.txt", 
-            $"Assets/Editor/Windows/{_windowFileName}/{_windowFileName}.uss"
-        );
-        
-        // Generate the UXML template.
-        ProjectWindowUtil.CreateScriptAssetFromTemplateFile(
-            "Packages/com.sebastian-inman.extended-editor-windows/Templates~/ExtendedEditorWindow.uxml.txt", 
-            $"Assets/Editor/Windows/{_windowFileName}/{_windowFileName}.uxml"
-        );
+        await FileGenerator.Generate($"{_filePath}.cs", new [] {
+            "using UnityEngine;",
+            "using ExtendedEditorWindows;",
+            "public class "+_windowFileName+" : ExtendedEditorWindow<"+_windowFileName+"> {",
+            "    [UnityEditor.MenuItem(\"" + _windowMenuPath + "\")]",
+            "    public static void Open() => OpenWindow(\""+_windowTitle+"\");",
+            "    protected override void Initialize() {",
+            "        Field(\"ExampleInput\", \"\", field => {",
+            "            Label(\"ExampleLabel\").text = field.value;",
+            "        });",
+            "    }",
+            "}"
+        });
 
-        AssetDatabase.Refresh();
+    }
+    
+    private async Task GenerateStyleFile() {
+        
+        await FileGenerator.Generate($"{_filePath}.uss", new [] {
+            "#ExampleLabel {",
+            "    color: dodgerblue;",
+            "    font-size: 18px;",
+            "}"
+        });
 
-        yield return null;
+    }
+    
+    private async Task GenerateTemplateFile() {
+        
+        await FileGenerator.Generate($"{_filePath}.uxml", new [] {
+            "<?xml version=\"1.0\" encoding=\"utf-8\"?>",
+            "<UXML xmlns=\"UnityEngine.UIElements\" xmlns:Editor=\"UnityEditor.UIElements\" class=\"h-full\">",
+            "    <VisualElement class=\"flex-1 flex flex-col\">",
+            "        <TextField name=\"ExampleInput\" label=\"Example Text Field\" class=\"mb-16\" />",
+            "        <Label name=\"ExampleLabel\" />",
+            "    </VisualElement>",
+            "</UXML>"
+        });
 
     }