+++ CHANGELOG.md
@@ -1,11 +1,22 @@
 ﻿# Changelog
 All notable changes to this project will be documented in this file.
 
-## [1.0.1] - 2022-02-24
+## [1.0.1] - 2022-02-25
+
+### Added
+
+- Added new `FileGenerator` helper class that asynchronously generates template files when creating a new Extended Editor Window.
+- Added a new tool for creating new Extended Editor Windows automatically. The tool can be found by going to `Assets > Create > Extended Editor Window`.
+
+## [1.0.0] - 2022-02-24
+
 ### Changed
+
 - Updated namespace from `I` to `ExtendedEditorWindow`
 - Helper classes now automatically derive the `rootVisualElement` element from the active editor window instead of requiring it as a parameter.
 
 ## [1.0.0] - 2022-02-24
+
 ### Added
+
 - First commit